feat(api-client): add composeSearchParameters URLSearchParams
The `composeSearchParameters` function is added to the `api-client` module. This function takes an object with a `pagination` property as an argument. If the `pagination` object contains a `limit` property, the function appends a `limit` parameter to the `URLSearchParams` object. If the `pagination` object contains an `offset` property, the function appends an `offset` parameter to the `URLSearchParams` object. If both `limit` and `offset` properties are present, both parameters are appended. If no pagination options are provided, the function returns `undefined`.