feat(publish-to-npm.yml): add workflow to publish moysklad-ts library to NPM
This commit adds a new file `.github/workflows/publish-to-npm.yml` which contains a workflow configuration to publish the moysklad-ts library to NPM. The workflow is triggered on a release event with the `published` type.

The workflow runs on the `ubuntu-latest` environment and consists of the following steps:
1. Checkout the repository using `actions/checkout@v4`.
2. Setup pnpm using `pnpm/action-setup@v2` with version 8.
3. Setup Node.js using `actions/setup-node@v3` with Node.js version 20, caching pnpm dependencies, and setting the registry URL to `https://registry.npmjs.org`.
4. Install dependencies using `pnpm install`.
5. Build the library using `pnpm build:lib`.
6. Publish the library to NPM using `pnpm publish --filter ./packages/moysklad-ts --access public` with the `NODE_AUTH_TOKEN` environment variable set to the value of the `NPM_TOKEN` secret.

This workflow enables automated publishing of the moysklad-ts library to NPM whenever a release is published.